+++ src/main/java/nextstep/security/filter/UserDetails.java
@@ -1,4 +1,4 @@
-package nextstep.security;
+package nextstep.security.filter;
 
 import java.util.Objects;
