+++ src/main/java/nextstep/security/filter/UserDetailsService.java
@@ -1,4 +1,4 @@
-package nextstep.security;
+package nextstep.security.filter;
 
 import java.util.Optional;
